refactor(layouts): restructure components into dedicated layout folders
Moved AuthLayout, MainLayout, and associated components into dedicated layout folders for better organization and maintainability. Updated import paths across the application to reflect the new structure.
