+++ src/App.jsx
@@ -1,15 +1,11 @@
 import './App.css'
-import Navbar from './components/Navbar'
-import Footer from './components/Footer'
 import Home from './pages/Home/Home'
 import Login from './pages/Login/Login'
 import { Route, Routes } from 'react-router-dom'
 import Register from './pages/Register/Register'
 import Landing from './pages/Landing/Landing'
 
-
 function App() {
-
   return (
     <>
       <Routes>
