+++ src/pages/Home/Home.jsx
@@ -1,21 +1,22 @@
-import Layout from '../../components/Layout'
-import { auth } from '../../firebase/config'
-import styles from './Home.module.css'
+import Layout from "../../layouts/MainLayout";
+import { auth } from "../../firebase/config";
+import styles from "./Home.module.css";
 
 const Home = () => {
   return (
     <Layout>
-      {auth.currentUser && <h1>Olá <span>{auth.currentUser.displayName}</span></h1> }
+      {auth.currentUser && (
+        <h1>
+          Olá <span>{auth.currentUser.displayName}</span>
+        </h1>
+      )}
       <h2>Veja os posts mais recentes</h2>
       <form className={styles.search_form}>
-        <input
-          type="text"
-          placeholder='Ou busque por tags...'
-        />
+        <input type="text" placeholder="Ou busque por tags..." />
         <button className="btn btn-dark">Pesquisar</button>
       </form>
     </Layout>
-  )
-}
+  );
+};
 
-export default Home
+export default Home;
